Chore: refactor rbl code into a pkg
