+++ pkg/rbl/rbl_test.go
@@ -1,10 +1,11 @@
-package collector
+package rbl_test
 
 import (
 	"os"
 	"testing"
 
 	"github.com/Luzilla/dnsbl_exporter/pkg/dns"
+	"github.com/Luzilla/dnsbl_exporter/pkg/rbl"
 	x "github.com/miekg/dns"
 	log "github.com/sirupsen/logrus"
 )
@@ -15,10 +16,10 @@ func TestUpdate(t *testing.T) {
 
 	d := dns.New(new(x.Client), "0.0.0.0:53")
 
-	rbl := NewRbl(d)
-	rbl.Update("this.is.not.an.ip", []string{"cbl.abuseat.org"})
+	r := rbl.New(d)
+	r.Update("this.is.not.an.ip", []string{"cbl.abuseat.org"})
 
-	if len(rbl.Results) > 0 {
-		t.Errorf("Got a result, but shouldn't have: %v", rbl.Results)
+	if len(r.Results) > 0 {
+		t.Errorf("Got a result, but shouldn't have: %v", r.Results)
 	}
 }
